+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-14T14:56:03.691Z",
+    "2026-07-14T16:25:29.466Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {