+++ packages/orq-rc/RELEASES.md
@@ -10799,3 +10799,13 @@ Based on:
 - [typescript v4.12.0-rc.24] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.24 - packages/orq-rc
+
+## 2026-07-14 16:28:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.1 (2.918.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.25] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.25] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.25 - packages/orq-rc